requirement.txt
annotated-types==0.7.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
djangorestframework==3.15.1
pydantic_core==2.18.4
sqlparse==0.5.0
typing_extensions==4.12.1
